distributed aspects, added router features, forwarding for servers

+++ internal/server/http.go
@@ -1,6 +1,7 @@
 package server
 
 import (
+	"distributed-kv-store/internal/ring"
 	"distributed-kv-store/internal/store"
 	"fmt"
 	"io"
@@ -8,15 +9,58 @@ import (
 )
 
 type Server struct {
-	Store *store.Store
+	Store  *store.Store
+	Router *Router
 }
 
-func NewServer(store *store.Store) *Server {
-	return &Server{Store: store}
+func NewServer(store *store.Store, self string, peers []string) *Server {
+	r := ring.NewRing()
+	r.AddNode(self)
+	for _, peer := range peers {
+		r.AddNode(peer)
+	}
+
+	return &Server{Store: store, Router: &Router{self: self, ring: r}}
+}
+
+func (s *Server) forward(w http.ResponseWriter, r *http.Request, owner string) {
+	url := "http://" + owner + r.URL.Path
+	req, err := http.NewRequest(r.Method, url, r.Body)
+	if err != nil {
+		http.Error(w, "forward error", http.StatusInternalServerError)
+		return
+	}
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		http.Error(w, "forward error", http.StatusBadGateway)
+		return
+	}
+	defer resp.Body.Close()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		http.Error(w, "read error", http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(resp.StatusCode)
+	w.Write(body)
 }
 
 func (s *Server) HandleGet(w http.ResponseWriter, r *http.Request) {
 	key := r.PathValue("key")
+	owner, err := s.Router.GetNode(key)
+	if err != nil {
+		http.Error(w, "ring error", http.StatusInternalServerError)
+		return
+	}
+
+	if owner != s.Router.self {
+		s.forward(w, r, owner)
+		return
+	}
+
 	val, ok := s.Store.Get(key)
 	if !ok {
 		http.Error(w, "key not found", http.StatusNotFound)
@@ -28,6 +72,18 @@ func (s *Server) HandleGet(w http.ResponseWriter, r *http.Request) {
 
 func (s *Server) HandlePut(w http.ResponseWriter, r *http.Request) {
 	key := r.PathValue("key")
+
+	owner, err := s.Router.GetNode(key)
+	if err != nil {
+		http.Error(w, "ring error", http.StatusInternalServerError)
+		return
+	}
+
+	if owner != s.Router.self {
+		s.forward(w, r, owner)
+		return
+	}
+
 	body, err := io.ReadAll(r.Body)
 	if err != nil {
 		http.Error(w, "body read error", http.StatusBadRequest)
@@ -35,6 +91,7 @@ func (s *Server) HandlePut(w http.ResponseWriter, r *http.Request) {
 	}
 
 	val := string(body)
+
 	err = s.Store.Put(key, val)
 	if err != nil {
 		http.Error(w, "value internal error", http.StatusInternalServerError)
@@ -46,6 +103,17 @@ func (s *Server) HandlePut(w http.ResponseWriter, r *http.Request) {
 
 func (s *Server) HandleDelete(w http.ResponseWriter, r *http.Request) {
 	key := r.PathValue("key")
+	owner, err := s.Router.GetNode(key)
+	if err != nil {
+		http.Error(w, "ring error", http.StatusInternalServerError)
+		return
+	}
+
+	if owner != s.Router.self {
+		s.forward(w, r, owner)
+		return
+	}
+
 	ok := s.Store.Delete(key)
 	if !ok {
 		http.Error(w, "delete key not found", http.StatusNotFound)
